+++ Composite/Composite/Ellipse.h
@@ -5,15 +5,15 @@
 class CEllipse : public CShape
 {
 public:
-	CEllipse(CPoint const& center, double horizontalRadius, double verticalRadius);
+	CEllipse(CPoint const& center, float horizontalRadius, float verticalRadius);
 	RectD GetFrame() override;
 	void SetFrame(const RectD & rect) override;
 protected:
 	void DrawBehavior(ICanvas & canvas)const override;
 private:
 	CPoint m_center;
-	double m_horizontalRadius;
-	double m_verticalRadius;
+	float m_horizontalRadius;
+	float m_verticalRadius;
 
 };
