+++ Composite/Composite/Triangle.cpp
@@ -11,11 +11,11 @@ CTriangle::CTriangle(CPoint const& point1, CPoint const& point2, CPoint const& p
 
 RectD CTriangle::GetFrame()
 {
-	double minX = std::min({ m_point1.GetX(), m_point1.GetX(), m_point1.GetX() });
-	double maxX = std::max({ m_point1.GetX(), m_point1.GetX(), m_point1.GetX() });
+	float minX = std::min({ m_point1.GetX(), m_point1.GetX(), m_point1.GetX() });
+	float maxX = std::max({ m_point1.GetX(), m_point1.GetX(), m_point1.GetX() });
 
-	double minY = std::min({ m_point1.GetY(), m_point1.GetY(), m_point1.GetY() });
-	double maxY = std::max({ m_point1.GetY(), m_point1.GetY(), m_point1.GetY() });
+	float minY = std::min({ m_point1.GetY(), m_point1.GetY(), m_point1.GetY() });
+	float maxY = std::max({ m_point1.GetY(), m_point1.GetY(), m_point1.GetY() });
 
 	return RectD{ minX, minY, maxX - minX, maxY - minY };
 }
@@ -24,8 +24,8 @@ void CTriangle::SetFrame(const RectD & rect)
 {
 	const auto oldFrame = GetFrame();
 	auto UpdatePoint = [&](CPoint & point) {
-		double scaleX = (point.GetX() - oldFrame.left) / oldFrame.width;
-		double scaleY = (point.GetY() - oldFrame.top) / oldFrame.height;
+		float scaleX = (point.GetX() - oldFrame.left) / oldFrame.width;
+		float scaleY = (point.GetY() - oldFrame.top) / oldFrame.height;
 		point = CPoint(rect.left + rect.left * scaleX, rect.top + rect.top * scaleY);
 	};
 	UpdatePoint(m_point1);
